+++ mail_base/__manifest__.py
@@ -1,10 +1,9 @@
-# -*- coding: utf-8 -*-
 {
     "name": "Mail Base",
     "summary": """Makes Mail extendable""",
     "category": "Discuss",
     "images": [],
-    "version": "10.0.1.0.5",
+    "version": "11.0.1.0.5",
 
     "author": "IT-Projects LLC, Pavel Romanchenko",
     "support": "apps@it-projects.info",