+++ mail_to/__manifest__.py
@@ -1,10 +1,9 @@
-# -*- coding: utf-8 -*-
 {
     "name": """Show message recipients""",
     "summary": """Allows you be sure, that all discussion participants were notified""",
     "category": "Discuss",
     "images": ['images/1.png'],
-    "version": "1.0.1",
+    "version": "11.0.1.0.1",
 
     "author": "IT-Projects LLC, Pavel Romanchenko",
     "support": "apps@it-projects.info",